requirements.txt
fastapi==0.115.8
google_api_python_client==2.154.0
langchain==0.3.17
langchain_community==0.3.16
langchain_openai==0.3.3
langchain_unstructured==0.1.6
langsmith==0.3.5
protobuf==5.29.3
pydantic==2.10.6
pymilvus==2.5.4
python-dotenv==1.0.1
Requests==2.32.3
streamlit==1.40.2
uvicorn